fix(rust): TDX deterministic ReportData binding (Month enum formatting)
verify_report_data() formatted the cert NotBefore as

            ts.year(), ts.month(), ts.day(), ts.hour(), ts.minute())

but ts.month() returns the time::Month enum whose Display impl prints
the English name ("May"). With {:02} the minimum width is shorter
than the value, so no padding is applied and the result was

    2026-May-01T17:18Z

instead of

    2026-05-01T17:18Z

The issuer (enclave-os-virtual/caddy/ratls) computes the binding from
NotBefore using Go's '2006-01-02T15:04Z' layout, so every TDX
deterministic verification mismatched ReportData with errors like:

  ReportData mismatch:
    got:      <issuer-computed digest>
    expected: <verifier-computed digest>

The bug was latent until the wallet's mandatory inspect-then-verify
flow (1.2.20+) actually exercised verify() against TDX targets.

Cast ts.month() to u8 so {:02} formats the numeric month.

+++ rust/src/ratls_client.rs
@@ -635,10 +635,14 @@ fn verify_report_data(der: &[u8], raw: &[u8], policy: &VerificationPolicy) -> Re
             // x509-parser's ASN1Time::to_rfc2822 gives us RFC 2822; we need ISO.
             // Use the raw datetime.
             let ts = nb.to_datetime();
+            // NOTE: `ts.month()` returns the `time::Month` enum whose Display
+            // impl prints the English name ("May"), so `{:02}` would yield
+            // "2026-May-01T17:18Z" and never match the issuer's binding.
+            // Cast to u8 to format as a zero-padded number.
             format!(
                 "{:04}-{:02}-{:02}T{:02}:{:02}Z",
                 ts.year(),
-                ts.month(),
+                ts.month() as u8,
                 ts.day(),
                 ts.hour(),
                 ts.minute()
